Added additional functionality for automation
Used the plt.switch_backend function to generate plots without showing, in order to save the numerous plots generated by the automated pipeline. This function is experimental and appears to throw an exception as the program is ending, but that should not be much of a problem for the reduction process.